requirements.txt
python-telegram-bot==21.5
anthropic==0.39.0
python-dotenv==1.0.1
tiktoken==0.8.0
pillow==11.0.0
